Nat/test/adding more tests

+++ apps/backend/src/event/event.service.ts
@@ -1,4 +1,9 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { PrismaService } from '@/prisma/prisma.service';
 import { Prisma } from '@prisma/client';
 import { ReqEventGetPublishedDto, ReqEventPostDraftDto, ReqEventPatchDto } from './event.schema';
@@ -12,6 +17,11 @@ export class EventService {
   ) {}
 
   async eventDelete(id: number, userId: number) {
+    const exists = await this.eventExists(id);
+
+    if (!exists) {
+      throw new NotFoundException(`Event with id ${id.toString()} not found`);
+    }
     try {
       return await this.prisma.event.delete({
         where: {
@@ -24,7 +34,7 @@ export class EventService {
         },
       });
     } catch {
-      throw new NotFoundException(`Event not found or no permission to delete it.`);
+      throw new UnauthorizedException(`No permission to delete event with id ${id.toString()}.`);
     }
   }
